Fix bad import of AUTH_USER_MODEL (#1)
The current mechanism used to load the user model in our fork of waffle
causes a lot of issues with circular dependency. The fix here addresses
the issue by not assumig that django would be ready by the time that
waffle is imported.

Our version of waffle is pretty ancient but the API has changed so it
would require extra refactor of our code base to bring it up to date.